Decode nested strings as String, not raw bytes
Always decode strings as bytes isn't very convenient. I have plans to
allow binary decoding of specified string parameters in the following
commit.

TorrentTest is disabled for a moment because it's impossible to decode
raw bytes value in a dictionary without data loss I'll extend the API
and rewrite the test accordingly soon.